docs build script update

+++ dev/build-docs.sh
@@ -1,3 +1,5 @@
 THISPATH=$(dirname "$0")
-rm -rf $THISPATH/../docs/_build
+if [ "-r" = $1 ]; then
+    rm -rf $THISPATH/../docs/_build
+fi
 sphinx-build -b html $THISPATH/../docs $THISPATH/../docs/_build
